Adds Mindtagger integration for error analysis (fixes #16)
which downloads Mindtagger (in latest DeepDive release), then creates a
task, and launches it in a user-specific port, and displays the GUI in
an iframe.

- `rels.open_mindtagger(num_sample=20, width='100%', height=600)`
- `rels.get_mindtagger_tags()`
- `rels.launch_mindtagger(task_name, item_generator, ...)`

As long as the GUI is running in the background, users can freely
add/update the tags, and get the tags to postprocess them from the rest
of the notebook cells.

Also updates early notebook cells to skip NLP markup steps by default if
there's cached result, and show some examples to give a sense of what
the data looks like.

Known issue: Mindtagger UI is not responsive enough for small screens,
so needs some touches to fit nicely into typical notebook widths and
heights.  However, this can be done independently from DDlite.