Pull parser and context modifications for Fonduer into contrib.
Want to have wrappers around the base objects so that we can
import from fonduer instead of snorkel. This commit also adds the
correct imports in fonduer's __init__.py to allow cleaner import
syntax from the Jupyter notebook.

This commit is current not functioning, and we need to debug the
SQL insert error of a non-null type.