Handle when candidates are passed in to lf_helpers.
Just like the textual/tabular lf_helpers, this updates the structural
lf_helpers to take a default behavior when a Candidate, rather than
a Span, is passed in. Rather than crashing, we usually pick the first,
last, or all Spans of a Candidate depending on the purpose of the
helper function.