Switch from distracting hover-zoom effect to medium-like zoom
The fact that images would slightly enlarge on hover was distracting
when scrolling through the blog post. This commit uses the medium-zoom
JS plugin [1] and added the data-action="zoom" to the images that
we want to zoom. I've comment out the hover effect code, rather than
removing it for now.

[1] https://github.com/francoischalifour/medium-zoom