Allow candidates to be n-tuples for n != 2
Adapting the intro tutorial for single entity recognition tasks
uncovered an assumption that every candidate had to be an 2-Tuple. If it
were an 1-Tuple, an Exception occurred, if it were an n-tuple the parts
indexed with i >= 2 were silently ignored.

This commit constructs the list of tuples by list comprehension. By
this, the arity of the tuple can be arbitrarily at that location. Note
however that other functions might expect lists of length two.