Fix bug of yielding empty spans
A bug occurs if the text of the span ends in one of the split tokens.
For example, "BC546-" will try to yield "BC546-", "BC546", and an empty
span with invalid char_start and char_end. This stops it from yielding
the empty span.

See https://github.com/HazyResearch/fonduer/pull/112/.

Co-authored-by: Hiromu Hota <hiromu.hota@hal.hitachi.com>